+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Added 16 colors of Sheep Coin Banks.
 - Added 3 types of Cow Coin Banks: Cow, Red Mooshroom, and Brown Mooshroom.
 - Added 2 types of Bee Coin Banks: Bee and Angry Bee.
+- Added 2 types of Guardian Coin Banks: Guardian and Elder Guardian.
 - Added Ravager Coin Bank.
 
 ## [1.0.1]